Using a service for cross component communication
