+++ src/app/app.component.ts
@@ -1,9 +1,7 @@
-import { Component, signal } from '@angular/core';
+import { Component } from '@angular/core';
 import { Header } from './header/header';
 import { UserInputComponent } from './user-input/user-input.component';
-import { InvestmentInput } from './user-input/investment-input';
 import { InvestmentResultsComponent } from './investment-results/investment-results.component';
-import { InvestmentResults } from './investment-results/investment-results-type';
 
 @Component({
   selector: 'app-root',
@@ -12,31 +10,4 @@ import { InvestmentResults } from './investment-results/investment-results-type'
   templateUrl: './app.component.html',
 })
 export class AppComponent {
-  resultsData = signal<InvestmentResults[] | undefined>(undefined);
-
-  onCalculateInvestmentResults(data: InvestmentInput) {
-    const { initialInvestment, annualInvestment, expectedReturn, duration } =
-      data;
-    const annualData = [];
-    let investmentValue = initialInvestment;
-
-    for (let i = 0; i < duration; i++) {
-      const year = i + 1;
-      const interestEarnedInYear = investmentValue * (expectedReturn / 100);
-      investmentValue += interestEarnedInYear + annualInvestment;
-      const totalInterest =
-        investmentValue - annualInvestment * year - initialInvestment;
-      annualData.push({
-        year: year,
-        interest: interestEarnedInYear,
-        valueEndOfYear: investmentValue,
-        annualInvestment: annualInvestment,
-        totalInterest: totalInterest,
-        totalAmountInvested: initialInvestment + annualInvestment * year,
-      });
-    }
-
-    this.resultsData.set(annualData);
-    // console.log(annualData);
-  }
 }
