+++ src/app/investment-results/investment-results.component.html
@@ -1,4 +1,4 @@
-@if (!results()) {
+@if (!results) {
 <p class="center">Please enter some values and press "Calculate</p>
 } @else {
 <table>
@@ -10,7 +10,7 @@
     <th>Investment Capital</th>
   </thead>
   <tbody>
-    @for (result of results(); track result.year){
+    @for (result of results; track result.year){
     <tr>
       <td>{{ result.year }}</td>
       <td>{{ result.valueEndOfYear | currency }}</td>
