+++ src/app/investment-results/investment-results.component.ts
@@ -1,6 +1,6 @@
-import { Component, input, Input } from '@angular/core';
-import { InvestmentResults } from './investment-results-type';
+import { Component, inject } from '@angular/core';
 import { CurrencyPipe } from '@angular/common';
+import { InvestmentService } from '../investment.service';
 
 @Component({
   selector: 'app-investment-results',
@@ -10,6 +10,13 @@ import { CurrencyPipe } from '@angular/common';
   styleUrl: './investment-results.component.css',
 })
 export class InvestmentResultsComponent {
-  results = input<InvestmentResults[]>()
-  // @Input() results?: InvestmentResults[];
+
+  //Alternatively
+  // constructor(private investmentService: InvestmentService) {}
+
+  private investmentService = inject(InvestmentService);
+
+  get results() {
+    return this.investmentService.resultData;
+  }
 }
