+++ src/app/investment.service.ts
@@ -1,0 +1,33 @@
+import { Injectable } from '@angular/core';
+import { InvestmentInput } from './user-input/investment-input';
+import { InvestmentResults } from './investment-results/investment-results-type';
+
+@Injectable({ providedIn: 'root' })
+export class InvestmentService {
+  resultData?: InvestmentResults[];
+
+  calculateInvestmentResults(data: InvestmentInput) {
+    const { initialInvestment, annualInvestment, expectedReturn, duration } =
+      data;
+    const annualData = [];
+    let investmentValue = initialInvestment;
+
+    for (let i = 0; i < duration; i++) {
+      const year = i + 1;
+      const interestEarnedInYear = investmentValue * (expectedReturn / 100);
+      investmentValue += interestEarnedInYear + annualInvestment;
+      const totalInterest =
+        investmentValue - annualInvestment * year - initialInvestment;
+      annualData.push({
+        year: year,
+        interest: interestEarnedInYear,
+        valueEndOfYear: investmentValue,
+        annualInvestment: annualInvestment,
+        totalInterest: totalInterest,
+        totalAmountInvested: initialInvestment + annualInvestment * year,
+      });
+    }
+
+    this.resultData = annualData;
+  }
+}
